// web/shared/api/types.gen.ts
export type AnimeSchema = {
    /**
     * Id
     */
    id: string;
    /**
     * Slug
     */
    slug: string;
    /**
     * Shikimori Id
     */
    shikimori_id?: number | null;
    /**
     * Age Rating
     */
    age_rating: number | null;
    /**
     * Title
     */
    title: string;
    /**
     * Description
     */
    description: string | null;
    /**
     * Poster
     */
    poster: string | null;
    /**
     * Tags
     */
    tags: Array<TagsScheme> | null;
    /**
     * Blocked Countries
     */
    blocked_countries: Array<string> | null;
    /**
     * Year
     */
    year: number | null;
    /**
     * Duration
     */
    duration: number | null;
    /**
     * Episodes Count
     */
    episodes_count: number | null;
    /**
     * Kinopoisk Rating
     */
    kinopoisk_rating: number | null;
    /**
     * Created At
     */
    created_at: string | null;
    /**
     * Updated At
     */
    updated_at: string | null;
    /**
     * Screenshots
     */
    screenshots: Array<string> | null;
    /**
     * Countries
     */
    countries: Array<string> | null;
    /**
     * Iframe Url
     */
    iframe_url?: string | null;
    /**
     * Type
     */
    type: string | null;
    /**
     * Anime Status
     */
    anime_status?: string | null;
    /**
     * Anime Studios
     */
    anime_studios?: Array<string> | null;
    /**
     * Translations
     */
    translations: Array<AnimeTranslationSchema> | null;
    /**
     * Shikimori Rating
     */
    shikimori_rating?: number | null;
};

/**
 * AnimeTranslationSchema
 */
export type AnimeTranslationSchema = {
    /**
     * Id
     */
    id: number;
    /**
     * Title
     */
    title: string;
    /**
     * Episodes Count
     */
    episodes_count: number | null;
};
// ...
export type TagsScheme = {
    /**
     * Id
     */
    id: number;
    /**
     * Genre
     */
    genre: string;
};